LPS-30443 fix issue: if a user can not see some entities in the control panel, when he tries to empty the recycle bin he would see an error message. Instead, he will just remove those entries that he is seeing. He will see the error message only if he can see an entry that he can not delete.

+++ portal-impl/src/com/liferay/portlet/trash/service/impl/TrashEntryServiceImpl.java
@@ -66,8 +66,21 @@ public class TrashEntryServiceImpl extends TrashEntryServiceBaseImpl {
 
 		List<TrashEntry> entries = trashEntryPersistence.findByGroupId(groupId);
 
+		PermissionChecker permissionChecker = getPermissionChecker();
+
 		for (TrashEntry entry : entries) {
 			try {
+				TrashHandler trashHandler =
+					TrashHandlerRegistryUtil.getTrashHandler(
+						entry.getClassName());
+
+				if (!trashHandler.hasTrashPermission(
+						permissionChecker, 0, entry.getClassPK(),
+						ActionKeys.VIEW)) {
+
+					continue;
+				}
+
 				deleteEntry(entry);
 			}
 			catch (TrashPermissionException tpe) {
